+++ libuv/src/unix/getaddrinfo.c
@@ -22,7 +22,7 @@
  * include any headers.
  */
 #ifndef _GNU_SOURCE
-# define _GNU_SOURCE
+#define _GNU_SOURCE
 #endif
 
 #include "uv.h"
@@ -38,65 +38,79 @@
 /* EAI_* constants. */
 #include <netdb.h>
 
-
 int uv__getaddrinfo_translate_error(int sys_err) {
   switch (sys_err) {
-  case 0: return 0;
+    case 0:
+      return 0;
 #if defined(EAI_ADDRFAMILY)
-  case EAI_ADDRFAMILY: return UV_EAI_ADDRFAMILY;
+    case EAI_ADDRFAMILY:
+      return UV_EAI_ADDRFAMILY;
 #endif
 #if defined(EAI_AGAIN)
-  case EAI_AGAIN: return UV_EAI_AGAIN;
+    case EAI_AGAIN:
+      return UV_EAI_AGAIN;
 #endif
 #if defined(EAI_BADFLAGS)
-  case EAI_BADFLAGS: return UV_EAI_BADFLAGS;
+    case EAI_BADFLAGS:
+      return UV_EAI_BADFLAGS;
 #endif
 #if defined(EAI_BADHINTS)
-  case EAI_BADHINTS: return UV_EAI_BADHINTS;
+    case EAI_BADHINTS:
+      return UV_EAI_BADHINTS;
 #endif
 #if defined(EAI_CANCELED)
-  case EAI_CANCELED: return UV_EAI_CANCELED;
+    case EAI_CANCELED:
+      return UV_EAI_CANCELED;
 #endif
 #if defined(EAI_FAIL)
-  case EAI_FAIL: return UV_EAI_FAIL;
+    case EAI_FAIL:
+      return UV_EAI_FAIL;
 #endif
 #if defined(EAI_FAMILY)
-  case EAI_FAMILY: return UV_EAI_FAMILY;
+    case EAI_FAMILY:
+      return UV_EAI_FAMILY;
 #endif
 #if defined(EAI_MEMORY)
-  case EAI_MEMORY: return UV_EAI_MEMORY;
+    case EAI_MEMORY:
+      return UV_EAI_MEMORY;
 #endif
 #if defined(EAI_NODATA)
-  case EAI_NODATA: return UV_EAI_NODATA;
+    case EAI_NODATA:
+      return UV_EAI_NODATA;
 #endif
 #if defined(EAI_NONAME)
-# if !defined(EAI_NODATA) || EAI_NODATA != EAI_NONAME
-  case EAI_NONAME: return UV_EAI_NONAME;
-# endif
+#if !defined(EAI_NODATA) || EAI_NODATA != EAI_NONAME
+    case EAI_NONAME:
+      return UV_EAI_NONAME;
+#endif
 #endif
 #if defined(EAI_OVERFLOW)
-  case EAI_OVERFLOW: return UV_EAI_OVERFLOW;
+    case EAI_OVERFLOW:
+      return UV_EAI_OVERFLOW;
 #endif
 #if defined(EAI_PROTOCOL)
-  case EAI_PROTOCOL: return UV_EAI_PROTOCOL;
+    case EAI_PROTOCOL:
+      return UV_EAI_PROTOCOL;
 #endif
 #if defined(EAI_SERVICE)
-  case EAI_SERVICE: return UV_EAI_SERVICE;
+    case EAI_SERVICE:
+      return UV_EAI_SERVICE;
 #endif
 #if defined(EAI_SOCKTYPE)
-  case EAI_SOCKTYPE: return UV_EAI_SOCKTYPE;
+    case EAI_SOCKTYPE:
+      return UV_EAI_SOCKTYPE;
 #endif
 #if defined(EAI_SYSTEM)
-  case EAI_SYSTEM: return UV__ERR(errno);
+    case EAI_SYSTEM:
+      return UV__ERR(errno);
 #endif
   }
   assert(!"unknown EAI_* error code");
   abort();
 #ifndef __SUNPRO_C
-  return 0;  /* Pacify compiler. */
+  return 0; /* Pacify compiler. */
 #endif
 }
-
 
 static void uv__getaddrinfo_work(struct uv__work* w) {
   uv_getaddrinfo_t* req;
@@ -106,7 +120,6 @@ static void uv__getaddrinfo_work(struct uv__work* w) {
   err = getaddrinfo(req->hostname, req->service, req->hints, &req->addrinfo);
   req->retcode = uv__getaddrinfo_translate_error(err);
 }
-
 
 static void uv__getaddrinfo_done(struct uv__work* w, int status) {
   uv_getaddrinfo_t* req;
@@ -137,13 +150,8 @@ static void uv__getaddrinfo_done(struct uv__work* w, int status) {
     req->cb(req, req->retcode, req->addrinfo);
 }
 
-
-int uv_getaddrinfo(uv_loop_t* loop,
-                   uv_getaddrinfo_t* req,
-                   uv_getaddrinfo_cb cb,
-                   const char* hostname,
-                   const char* service,
-                   const struct addrinfo* hints) {
+int uv_getaddrinfo(uv_loop_t* loop, uv_getaddrinfo_t* req, uv_getaddrinfo_cb cb, const char* hostname,
+                   const char* service, const struct addrinfo* hints) {
   char hostname_ascii[256];
   size_t hostname_len;
   size_t service_len;
@@ -155,19 +163,17 @@ int uv_getaddrinfo(uv_loop_t* loop,
   if (req == NULL || (hostname == NULL && service == NULL))
     return UV_EINVAL;
 
-  /* FIXME(bnoordhuis) IDNA does not seem to work z/OS,
-   * probably because it uses EBCDIC rather than ASCII.
-   */
+    /* FIXME(bnoordhuis) IDNA does not seem to work z/OS,
+     * probably because it uses EBCDIC rather than ASCII.
+     */
 #ifdef __MVS__
-  (void) &hostname_ascii;
+  (void)&hostname_ascii;
 #else
   if (hostname != NULL) {
-    rc = uv__idna_toascii(hostname,
-                          hostname + strlen(hostname),
-                          hostname_ascii,
-                          hostname_ascii + sizeof(hostname_ascii));
+    rc = uv__idna_toascii(
+        hostname, hostname + strlen(hostname), hostname_ascii, hostname_ascii + sizeof(hostname_ascii));
     if (rc < 0)
-      return rc;
+      return (int)rc;
     hostname = hostname_ascii;
   }
 #endif
@@ -206,11 +212,7 @@ int uv_getaddrinfo(uv_loop_t* loop,
     req->hostname = memcpy(buf + len, hostname, hostname_len);
 
   if (cb) {
-    uv__work_submit(loop,
-                    &req->work_req,
-                    UV__WORK_SLOW_IO,
-                    uv__getaddrinfo_work,
-                    uv__getaddrinfo_done);
+    uv__work_submit(loop, &req->work_req, UV__WORK_SLOW_IO, uv__getaddrinfo_work, uv__getaddrinfo_done);
     return 0;
   } else {
     uv__getaddrinfo_work(&req->work_req);
@@ -219,12 +221,10 @@ int uv_getaddrinfo(uv_loop_t* loop,
   }
 }
 
-
 void uv_freeaddrinfo(struct addrinfo* ai) {
   if (ai)
     freeaddrinfo(ai);
 }
-
 
 int uv_if_indextoname(unsigned int ifindex, char* buffer, size_t* size) {
   char ifname_buf[UV_IF_NAMESIZE];
