+++ libuv/src/unix/macosx/darwin.c
@@ -30,8 +30,7 @@
 #include <mach-o/dyld.h> /* _NSGetExecutablePath */
 #include <sys/resource.h>
 #include <sys/sysctl.h>
-#include <unistd.h>  /* sysconf */
-
+#include <unistd.h> /* sysconf */
 
 int uv__platform_loop_init(uv_loop_t* loop) {
   loop->cf_state = NULL;
@@ -42,23 +41,19 @@ int uv__platform_loop_init(uv_loop_t* loop) {
   return 0;
 }
 
-
 void uv__platform_loop_delete(uv_loop_t* loop) {
   uv__fsevents_loop_delete(loop);
 }
 
-
 uint64_t uv__hrtime(uv_clocktype_t type) {
   static mach_timebase_info_data_t info;
 
-  if ((ACCESS_ONCE(uint32_t, info.numer) == 0 ||
-       ACCESS_ONCE(uint32_t, info.denom) == 0) &&
+  if ((ACCESS_ONCE(uint32_t, info.numer) == 0 || ACCESS_ONCE(uint32_t, info.denom) == 0) &&
       mach_timebase_info(&info) != KERN_SUCCESS)
     abort();
 
   return mach_absolute_time() * info.numer / info.denom;
 }
-
 
 int uv_exepath(char* buffer, size_t* size) {
   /* realpath(exepath) may be > PATH_MAX so double it to be on the safe side. */
@@ -91,19 +86,16 @@ int uv_exepath(char* buffer, size_t* size) {
   return 0;
 }
 
-
 uint64_t uv_get_free_memory(void) {
   vm_statistics_data_t info;
   mach_msg_type_number_t count = sizeof(info) / sizeof(integer_t);
 
-  if (host_statistics(mach_host_self(), HOST_VM_INFO,
-                      (host_info_t)&info, &count) != KERN_SUCCESS) {
-    return UV_EINVAL;  /* FIXME(bnoordhuis) Translate error. */
+  if (host_statistics(mach_host_self(), HOST_VM_INFO, (host_info_t)&info, &count) != KERN_SUCCESS) {
+    return UV_EINVAL; /* FIXME(bnoordhuis) Translate error. */
   }
 
-  return (uint64_t) info.free_count * sysconf(_SC_PAGESIZE);
+  return (uint64_t)info.free_count * sysconf(_SC_PAGESIZE);
 }
-
 
 uint64_t uv_get_total_memory(void) {
   uint64_t info;
@@ -113,27 +105,25 @@ uint64_t uv_get_total_memory(void) {
   if (sysctl(which, 2, &info, &size, NULL, 0))
     return UV__ERR(errno);
 
-  return (uint64_t) info;
+  return (uint64_t)info;
 }
-
 
 uint64_t uv_get_constrained_memory(void) {
-  return 0;  /* Memory constraints are unknown. */
+  return 0; /* Memory constraints are unknown. */
 }
-
 
 void uv_loadavg(double avg[3]) {
   struct loadavg info;
   size_t size = sizeof(info);
   int which[] = {CTL_VM, VM_LOADAVG};
 
-  if (sysctl(which, 2, &info, &size, NULL, 0) < 0) return;
+  if (sysctl(which, 2, &info, &size, NULL, 0) < 0)
+    return;
 
-  avg[0] = (double) info.ldavg[0] / info.fscale;
-  avg[1] = (double) info.ldavg[1] / info.fscale;
-  avg[2] = (double) info.ldavg[2] / info.fscale;
+  avg[0] = (double)info.ldavg[0] / info.fscale;
+  avg[1] = (double)info.ldavg[1] / info.fscale;
+  avg[2] = (double)info.ldavg[2] / info.fscale;
 }
-
 
 int uv_resident_set_memory(size_t* rss) {
   mach_msg_type_number_t count;
@@ -141,11 +131,8 @@ int uv_resident_set_memory(size_t* rss) {
   kern_return_t err;
 
   count = TASK_BASIC_INFO_COUNT;
-  err = task_info(mach_task_self(),
-                  TASK_BASIC_INFO,
-                  (task_info_t) &info,
-                  &count);
-  (void) &err;
+  err = task_info(mach_task_self(), TASK_BASIC_INFO, (task_info_t)&info, &count);
+  (void)&err;
   /* task_info(TASK_BASIC_INFO) cannot really fail. Anything other than
    * KERN_SUCCESS implies a libuv bug.
    */
@@ -154,7 +141,6 @@ int uv_resident_set_memory(size_t* rss) {
 
   return 0;
 }
-
 
 int uv_uptime(double* uptime) {
   time_t now;
@@ -172,15 +158,14 @@ int uv_uptime(double* uptime) {
 }
 
 int uv_cpu_info(uv_cpu_info_t** cpu_infos, int* count) {
-  unsigned int ticks = (unsigned int)sysconf(_SC_CLK_TCK),
-               multiplier = ((uint64_t)1000L / ticks);
+  unsigned int ticks = (unsigned int)sysconf(_SC_CLK_TCK), multiplier = ((uint64_t)1000L / ticks);
   char model[512];
   uint64_t cpuspeed;
   size_t size;
   unsigned int i;
   natural_t numcpus;
   mach_msg_type_number_t msg_type;
-  processor_cpu_load_info_data_t *info;
+  processor_cpu_load_info_data_t* info;
   uv_cpu_info_t* cpu_info;
 
   size = sizeof(model);
@@ -193,10 +178,10 @@ int uv_cpu_info(uv_cpu_info_t** cpu_infos, int* count) {
   if (sysctlbyname("hw.cpufrequency", &cpuspeed, &size, NULL, 0))
     return UV__ERR(errno);
 
-  if (host_processor_info(mach_host_self(), PROCESSOR_CPU_LOAD_INFO, &numcpus,
-                          (processor_info_array_t*)&info,
-                          &msg_type) != KERN_SUCCESS) {
-    return UV_EINVAL;  /* FIXME(bnoordhuis) Translate error. */
+  if (host_processor_info(
+          mach_host_self(), PROCESSOR_CPU_LOAD_INFO, &numcpus, (processor_info_array_t*)&info, &msg_type) !=
+      KERN_SUCCESS) {
+    return UV_EINVAL; /* FIXME(bnoordhuis) Translate error. */
   }
 
   *cpu_infos = uv__malloc(numcpus * sizeof(**cpu_infos));
@@ -217,13 +202,12 @@ int uv_cpu_info(uv_cpu_info_t** cpu_infos, int* count) {
     cpu_info->cpu_times.irq = 0;
 
     cpu_info->model = uv__strdup(model);
-    cpu_info->speed = cpuspeed/1000000;
+    cpu_info->speed = (int)(cpuspeed / 1000000);
   }
   vm_deallocate(mach_task_self(), (vm_address_t)info, msg_type);
 
   return 0;
 }
-
 
 void uv_free_cpu_info(uv_cpu_info_t* cpu_infos, int count) {
   int i;
