+++ libuv/src/unix/macosx/kqueue.c
@@ -42,11 +42,10 @@
  * http://www.boost.org/doc/libs/1_61_0/boost/asio/detail/kqueue_reactor.hpp
  */
 #ifndef EV_OOBAND
-#define EV_OOBAND  EV_FLAG1
+#define EV_OOBAND EV_FLAG1
 #endif
 
 static void uv__fs_event(uv_loop_t* loop, uv__io_t* w, unsigned int fflags);
-
 
 int uv__kqueue_init(uv_loop_t* loop) {
   loop->backend_fd = kqueue();
@@ -57,7 +56,6 @@ int uv__kqueue_init(uv_loop_t* loop) {
 
   return 0;
 }
-
 
 #if defined(__APPLE__) && MAC_OS_X_VERSION_MAX_ALLOWED >= 1070
 static int uv__has_forked_with_cfrunloop;
@@ -90,7 +88,6 @@ int uv__io_fork(uv_loop_t* loop) {
   return err;
 }
 
-
 int uv__io_check_fd(uv_loop_t* loop, int fd) {
   struct kevent ev;
   int rc;
@@ -107,7 +104,6 @@ int uv__io_check_fd(uv_loop_t* loop, int fd) {
 
   return rc;
 }
-
 
 void uv__io_poll(uv_loop_t* loop, int timeout) {
   struct kevent events[1024];
@@ -145,7 +141,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
     w = QUEUE_DATA(q, uv__io_t, watcher_queue);
     assert(w->pevents != 0);
     assert(w->fd >= 0);
-    assert(w->fd < (int) loop->nwatchers);
+    assert(w->fd < (int)loop->nwatchers);
 
     if ((w->events & POLLIN) == 0 && (w->pevents & POLLIN) != 0) {
       filter = EVFILT_READ;
@@ -154,8 +150,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
 
       if (w->cb == uv__fs_event) {
         filter = EVFILT_VNODE;
-        fflags = NOTE_ATTRIB | NOTE_WRITE  | NOTE_RENAME
-               | NOTE_DELETE | NOTE_EXTEND | NOTE_REVOKE;
+        fflags = NOTE_ATTRIB | NOTE_WRITE | NOTE_RENAME | NOTE_DELETE | NOTE_EXTEND | NOTE_REVOKE;
         op = EV_ADD | EV_ONESHOT; /* Stop the event from firing repeatedly. */
       }
 
@@ -178,7 +173,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
       }
     }
 
-   if ((w->events & UV__POLLPRI) == 0 && (w->pevents & UV__POLLPRI) != 0) {
+    if ((w->events & UV__POLLPRI) == 0 && (w->pevents & UV__POLLPRI) != 0) {
       EV_SET(events + nevents, w->fd, EV_OOBAND, EV_ADD, 0, 0, 0);
 
       if (++nevents == ARRAY_SIZE(events)) {
@@ -211,12 +206,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
     if (pset != NULL)
       pthread_sigmask(SIG_BLOCK, pset, NULL);
 
-    nfds = kevent(loop->backend_fd,
-                  events,
-                  nevents,
-                  events,
-                  ARRAY_SIZE(events),
-                  timeout == -1 ? NULL : &spec);
+    nfds = kevent(loop->backend_fd, events, nevents, events, ARRAY_SIZE(events), timeout == -1 ? NULL : &spec);
 
     if (pset != NULL)
       pthread_sigmask(SIG_UNBLOCK, pset, NULL);
@@ -250,11 +240,11 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
     nevents = 0;
 
     assert(loop->watchers != NULL);
-    loop->watchers[loop->nwatchers] = (void*) events;
-    loop->watchers[loop->nwatchers + 1] = (void*) (uintptr_t) nfds;
+    loop->watchers[loop->nwatchers] = (void*)events;
+    loop->watchers[loop->nwatchers + 1] = (void*)(uintptr_t)nfds;
     for (i = 0; i < nfds; i++) {
       ev = events + i;
-      fd = ev->ident;
+      fd = (int)ev->ident;
       /* Skip invalidated events, see uv__platform_invalidate_fd */
       if (fd == -1)
         continue;
@@ -286,7 +276,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
       if (ev->filter == EVFILT_READ) {
         if (w->pevents & POLLIN) {
           revents |= POLLIN;
-          w->rcount = ev->data;
+          w->rcount = (int)ev->data;
         } else {
           /* TODO batch up */
           struct kevent events[1];
@@ -300,7 +290,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
       if (ev->filter == EV_OOBAND) {
         if (w->pevents & UV__POLLPRI) {
           revents |= UV__POLLPRI;
-          w->rcount = ev->data;
+          w->rcount = (int)ev->data;
         } else {
           /* TODO batch up */
           struct kevent events[1];
@@ -314,7 +304,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
       if (ev->filter == EVFILT_WRITE) {
         if (w->pevents & POLLOUT) {
           revents |= POLLOUT;
-          w->wcount = ev->data;
+          w->wcount = (int)ev->data;
         } else {
           /* TODO batch up */
           struct kevent events[1];
@@ -352,7 +342,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
     loop->watchers[loop->nwatchers + 1] = NULL;
 
     if (have_signals != 0)
-      return;  /* Event loop should cycle now so don't poll again. */
+      return; /* Event loop should cycle now so don't poll again. */
 
     if (nevents != 0) {
       if (nfds == ARRAY_SIZE(events) && --count != 0) {
@@ -369,17 +359,16 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
     if (timeout == -1)
       continue;
 
-update_timeout:
+  update_timeout:
     assert(timeout > 0);
 
     diff = loop->time - base;
-    if (diff >= (uint64_t) timeout)
+    if (diff >= (uint64_t)timeout)
       return;
 
     timeout -= diff;
   }
 }
-
 
 void uv__platform_invalidate_fd(uv_loop_t* loop, int fd) {
   struct kevent* events;
@@ -389,17 +378,16 @@ void uv__platform_invalidate_fd(uv_loop_t* loop, int fd) {
   assert(loop->watchers != NULL);
   assert(fd >= 0);
 
-  events = (struct kevent*) loop->watchers[loop->nwatchers];
-  nfds = (uintptr_t) loop->watchers[loop->nwatchers + 1];
+  events = (struct kevent*)loop->watchers[loop->nwatchers];
+  nfds = (uintptr_t)loop->watchers[loop->nwatchers + 1];
   if (events == NULL)
     return;
 
   /* Invalidate events with same file descriptor */
   for (i = 0; i < nfds; i++)
-    if ((int) events[i].ident == fd)
+    if ((int)events[i].ident == fd)
       events[i].ident = -1;
 }
-
 
 static void uv__fs_event(uv_loop_t* loop, uv__io_t* w, unsigned int fflags) {
   uv_fs_event_t* handle;
@@ -433,8 +421,7 @@ static void uv__fs_event(uv_loop_t* loop, uv__io_t* w, unsigned int fflags) {
     return;
 
   /* Watcher operates in one-shot mode, re-arm it. */
-  fflags = NOTE_ATTRIB | NOTE_WRITE  | NOTE_RENAME
-         | NOTE_DELETE | NOTE_EXTEND | NOTE_REVOKE;
+  fflags = NOTE_ATTRIB | NOTE_WRITE | NOTE_RENAME | NOTE_DELETE | NOTE_EXTEND | NOTE_REVOKE;
 
   EV_SET(&ev, w->fd, EVFILT_VNODE, EV_ADD | EV_ONESHOT, fflags, 0, 0);
 
@@ -442,17 +429,12 @@ static void uv__fs_event(uv_loop_t* loop, uv__io_t* w, unsigned int fflags) {
     abort();
 }
 
-
 int uv_fs_event_init(uv_loop_t* loop, uv_fs_event_t* handle) {
   uv__handle_init(loop, (uv_handle_t*)handle, UV_FS_EVENT);
   return 0;
 }
 
-
-int uv_fs_event_start(uv_fs_event_t* handle,
-                      uv_fs_event_cb cb,
-                      const char* path,
-                      unsigned int flags) {
+int uv_fs_event_start(uv_fs_event_t* handle, uv_fs_event_cb cb, const char* path, unsigned int flags) {
   int fd;
 
   if (uv__is_active(handle))
@@ -503,7 +485,6 @@ int uv_fs_event_start(uv_fs_event_t* handle,
   return 0;
 }
 
-
 int uv_fs_event_stop(uv_fs_event_t* handle) {
   int r;
   r = 0;
@@ -529,7 +510,6 @@ int uv_fs_event_stop(uv_fs_event_t* handle) {
 
   return r;
 }
-
 
 void uv__fs_event_close(uv_fs_event_t* handle) {
   uv_fs_event_stop(handle);
